refactor: utils 폴더 네이밍 컨벤션 통일 (*.util.ts)
