+++ src/lib/api/estimate/requests/createDesignatedEstimate.ts
@@ -1,11 +1,14 @@
-import { tokenFetch } from "@/lib/utils/fetch-client";
+import { tokenFetch } from "@/lib/utils";
 
-export async function createDesignatedEstimate(moverId: string, requestId: string) {
-  return await tokenFetch(`/requests/movers/${moverId}`, {
-    method: 'PATCH',
-    body: JSON.stringify({ requestId }),
-    headers: {
-      'Content-Type': 'application/json',
-    },
-  });
+export async function createDesignatedEstimate(
+   moverId: string,
+   requestId: string,
+) {
+   return await tokenFetch(`/requests/movers/${moverId}`, {
+      method: "PATCH",
+      body: JSON.stringify({ requestId }),
+      headers: {
+         "Content-Type": "application/json",
+      },
+   });
 }
