+++ src/lib/api/mover/requests/favoriteMover.ts
@@ -1,11 +1,11 @@
-import { tokenFetch } from "@/lib/utils/fetch-client";
+import { tokenFetch } from "@/lib/utils";
 
 export const toggleFavoriteMover = async (moverId: string) => {
-  return await tokenFetch(`/movers/${moverId}/toggle-favorite`, {
-    method: "POST",
-    body: JSON.stringify({}),
-    headers: {
-      'Content-Type': 'application/json',
-    },
-  });
+   return await tokenFetch(`/movers/${moverId}/toggle-favorite`, {
+      method: "POST",
+      body: JSON.stringify({}),
+      headers: {
+         "Content-Type": "application/json",
+      },
+   });
 };
