+++ src/lib/api/mover/requests/getMover.ts
@@ -1,5 +1,5 @@
 //lib/api/mover/getMover.ts
-import { tokenFetch, defaultFetch } from "@/lib/utils/fetch-client";
+import { defaultFetch, tokenFetch } from "@/lib/utils";
 import { delay } from "../../../../../delay";
 import { GetMoversParams, GetMoversResponse, Mover } from "@/lib/types";
