+++ src/lib/api/review/requests/createReview.ts
@@ -1,5 +1,5 @@
 import { CreateReviewDto } from "@/lib/schemas";
-import { tokenFetch } from "@/lib/utils/fetch-client";
+import { tokenFetch } from "@/lib/utils";
 
 export async function createReview(data: CreateReviewDto) {
    return await tokenFetch(`/reviews`, {
