+++ src/lib/utils/ai.util.ts
