+++ src/lib/utils/index.ts
@@ -1,13 +1,12 @@
-export * from "./date.utils";
 export * from "./address.util";
+export * from "./ai.util";
 export * from "./auth.util";
-export * from "./fetch-client";
-export * from "./fetch-error.util";
-export * from "./moveChip.util";
-export * from "./extractErrorMessage.utill";
-export * from "./middleware.util";
-export * from "./formatAddress.util";
-export * from "./price.utils";
-export * from "./ai.utils";
-export * from "./distance.utils";
 export * from "./chatNotification.util";
+export * from "./date.util";
+export * from "./distance.util";
+export * from "./extractErrorMessage.util";
+export * from "./fetchClient.util";
+export * from "./fetchError.util";
+export * from "./formatAddress.util";
+export * from "./middleware.util";
+export * from "./moveChip.util";
